Replace stat call with NSFileManager API on Apple target
Starting from May 1, 2024 developers have to explicitly
declare why they use APIs allowing to access
file timestamps in a privacy manifest files;
otherwise an app won't be accepted by the AppStore.
The only restricted API we're currently using
is stat, and since we don't extract any timestamps
using it, we can safely replace it with
NSFileManager::fileAttributesAtPath.

For more details on the restriction imposed on
timestamp-accessing APIs read https://developer.apple.com/documentation/bundleresources/privacy_manifest_files/describing_use_of_required_reason_api

Closes #297
