+++ core/native/src/files/FileSystemNative.kt
@@ -63,23 +63,7 @@ public actual val SystemFileSystem: FileSystem = object : SystemFileSystemImpl()
         atomicMoveImpl(source, destination)
     }
 
-    @OptIn(ExperimentalForeignApi::class, UnsafeNumber::class)
-    override fun metadataOrNull(path: Path): FileMetadata? {
-        memScoped {
-            val struct_stat = alloc<stat>()
-            if (stat(path.path, struct_stat.ptr) != 0) {
-                if (errno == ENOENT) return null
-                throw IOException("stat failed to ${path.path}: ${strerror(errno)?.toKString()}")
-            }
-            val mode = struct_stat.st_mode.toInt()
-            val isFile = (mode and S_IFMT) == S_IFREG
-            return FileMetadata(
-                isRegularFile = isFile,
-                isDirectory = (mode and S_IFMT) == S_IFDIR,
-                if (isFile) struct_stat.st_size.toLong() else -1L
-            )
-        }
-    }
+    override fun metadataOrNull(path: Path): FileMetadata? = metadataOrNullImpl(path)
 
     override fun resolve(path: Path): Path {
         if (!exists(path)) throw FileNotFoundException(path.path)
@@ -103,6 +87,8 @@ public actual val SystemFileSystem: FileSystem = object : SystemFileSystemImpl()
         return FileSink(openFile)
     }
 }
+
+internal expect fun metadataOrNullImpl(path: Path): FileMetadata?
 
 internal expect fun atomicMoveImpl(source: Path, destination: Path)
